Fix Bolan ducking y position bug
It was previously necessary to move Bolan's y-position down to make
sure he stays on the floor while ducking. This caused an animation bug
wherein he would be suspended in a running sprite for a single frame
before his sprite switched.

The solution was to leave the whitespace in the ducking sprites.